+++ docs/src/pages/starters/gatsby.js
@@ -3,11 +3,12 @@ import Layout from '@theme/Layout';
 import StarterComponentPage from '../sections/StarterComponentPage/StarterComponentPage'
 
 const data = {
-  name: 'Base Store',
+  name: 'Gatsby Store',
   owner: 'VTEX',
   description:
     'Kickoff your store with this boilerplate. This starter ships with the main FastStore configuration files you need to get up and running blazing fast.',
   features: [
+    'Gatsby',
     'Landing page',
     'Product page',
     'Intelligent search',
@@ -15,15 +16,15 @@ const data = {
     'Optimistic shopping cart',
     'Google Analytics',
   ],
-  demoURL: 'https://base.vtex.app/',
-  gitHubURL: 'https://github.com/vtex-sites/base.store',
+  demoURL: 'https://gatsby.vtex.app/',
+  gitHubURL: 'https://github.com/vtex-sites/gatsby.store',
   price: 'Free',
   img: '/img/starters/base-starter.png',
 }
 
 function STARTERS() {
   return (
-    <Layout title="Base Starter">
+    <Layout title="Gatsby Starter">
       <StarterComponentPage data={data} />
     </Layout>
   );