+++ packages/api/src/platforms/vtex/clients/index.ts
@@ -1,5 +1,6 @@
 import { VtexCommerce } from './commerce'
 import { IntelligentSearch } from './search'
+import { SP } from './sp'
 import type { Context, Options } from '..'
 
 export type Clients = ReturnType<typeof getClients>
@@ -7,9 +8,11 @@ export type Clients = ReturnType<typeof getClients>
 export const getClients = (options: Options, ctx: Context) => {
   const search = IntelligentSearch(options, ctx)
   const commerce = VtexCommerce(options, ctx)
+  const sp = SP(options, ctx)
 
   return {
     search,
     commerce,
+    sp,
   }
 }